계산 로직 CalculateProcessor로 이동
